backend/python/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.23.2
pydantic>=1.10.7
requests>=2.28.0
beautifulsoup4>=4.11.0
smolagents==1.13.0
python-dotenv>=1.0.0
